Packet capture and SIP updates, layout polish, Windows vendor libs
- Rust: database, user agent, DSCP, capture/pcap pipeline and SIP transport
- Frontend: shell layout, packet monitor, Text Forge, styles and scripts
- Vendor: x86_64 Windows Npcap/Vosk libs and manifest; Tauri windows schema
- Ignore .tmp for local SDK extractions

Made-with: Cursor

+++ src-tauri/src/packet_capture/pcap_reader.rs
@@ -224,7 +224,7 @@ impl MmapPcapReader {
 
         let pcap_header = pcap::PacketHeader {
             ts: libc::timeval {
-                tv_sec: header.ts_sec as i64,
+                tv_sec: header.ts_sec as i32,
                 tv_usec: ts_usec as i32,
             },
             caplen: header.caplen,
